Refactor: Split services and routes by feature
