+++ docs/DEVELOPMENT.md
@@ -349,6 +349,7 @@ if (!reportTypeAvailable(report)) {
 ### File Organization
 
 - Source code: `src/`
+- Feature services and route handlers: `src/features/*/service.ts` and `src/features/*/routes.ts` (`src/services.ts` and `src/routes.ts` remain temporary compatibility entrypoints during the reorganization)
 - Tests: `src/__tests__/`
 - Database layer: `src/db/`
 - CSV/data mappings: `src/features/stats/mapping.ts` (`src/mappings.ts` remains a temporary compatibility re-export used by legacy imports and scripts)
